Toast component files layout refactoring
