+++ .storybook/config.js
@@ -3,6 +3,7 @@ import { configure } from '@storybook/react';
 // automatically import all files ending in *.stories.js
 const reqRoutes = require.context('../src/routes', true, /\.stories\.((js|ts)x?)$/);
 const reqComponents = require.context('../src/components', true, /\.stories\.((js|ts)x?)$/);
+const ctxComponents = require.context('../src/context', true, /\.stories\.((js|ts)x?)$/);
 const req = require.context('../stories', true, /.stories.js$/);
 
 function loadStories() {
