+++ src/index.tsx
@@ -5,8 +5,8 @@ import * as ReactDOM from "react-dom";
 import { hot } from "react-hot-loader/root";
 import { Provider } from "react-redux";
 import WebFont from "webfontloader";
-import { ToastProvider } from "./components/layout/ToastProvider";
 import MatchMedia from "./context/MatchMediaContext";
+import { ToastProvider } from "./context/ToastProvider";
 import Route from "./routes";
 import { history, makeStore } from "./store";
 import theme from "./theme/mui";
